extract alert read models service
