+++ apps/api/src/alerting.ts
@@ -1,6 +1,7 @@
 export { ANTICIPATED_ERROR_IDENTIFIERS } from "@maple/domain/anticipated-errors"
 export { AlertRuntime, AlertsService } from "./services/alerts/AlertsService"
 export { AlertDestinationsService } from "./services/alerts/AlertDestinationsService"
+export { AlertReadModelsService } from "./services/alerts/AlertReadModelsService"
 export { AnomalyDetectionService } from "./services/alerts/AnomalyDetectionService"
 export { BucketCacheService } from "@maple/query-engine/caching"
 export { CacheBackendLive } from "@/platform/CacheBackendLive"
