+++ apps/api/src/mcp/tools/get-incident-timeline.ts
@@ -3,7 +3,7 @@ import { formatNextSteps } from "@/mcp/lib/next-steps"
 import { Effect, Schema } from "effect"
 import { createDualContent } from "@/mcp/lib/structured-output"
 import { resolveTenant } from "@/mcp/lib/query-warehouse"
-import { AlertsService } from "@/services/alerts/AlertsService"
+import { AlertReadModelsService } from "@/services/alerts/AlertReadModelsService"
 
 const comparatorLabel: Record<string, string> = {
 	gt: ">",
@@ -42,9 +42,9 @@ export function registerGetIncidentTimelineTool(server: McpToolRegistrar) {
 		}),
 		Effect.fn("McpTool.getIncidentTimeline")(function* ({ rule_id, status, severity, group_key, limit }) {
 			const tenant = yield* resolveTenant
-			const alerts = yield* AlertsService
+			const readModels = yield* AlertReadModelsService
 
-			const result = yield* alerts.listIncidents(tenant.orgId).pipe(
+			const result = yield* readModels.listIncidents(tenant.orgId).pipe(
 				Effect.mapError(
 					(error) =>
 						new McpQueryError({
