+++ apps/api/src/mcp/tools/list-alert-incidents.ts
@@ -4,7 +4,7 @@ import { formatNextSteps } from "@/mcp/lib/next-steps"
 import { Effect, Schema } from "effect"
 import { createDualContent } from "@/mcp/lib/structured-output"
 import { resolveTenant } from "@/mcp/lib/query-warehouse"
-import { AlertsService } from "@/services/alerts/AlertsService"
+import { AlertReadModelsService } from "@/services/alerts/AlertReadModelsService"
 
 const comparatorLabel: Record<string, string> = {
 	gt: ">",
@@ -25,9 +25,9 @@ export function registerListAlertIncidentsTool(server: McpToolRegistrar) {
 		}),
 		Effect.fn("McpTool.listAlertIncidents")(function* ({ status, severity, group_key, limit }) {
 			const tenant = yield* resolveTenant
-			const alerts = yield* AlertsService
+			const readModels = yield* AlertReadModelsService
 
-			const result = yield* alerts.listIncidents(tenant.orgId).pipe(
+			const result = yield* readModels.listIncidents(tenant.orgId).pipe(
 				Effect.mapError(
 					(error) =>
 						new McpQueryError({
