+++ apps/api/src/routes/v2/alchemy-provider.integration.test.ts
@@ -35,6 +35,7 @@ import { AuthService } from "@/services/auth/AuthService"
 import { DashboardPersistenceService } from "@/services/dashboards/DashboardPersistenceService"
 import { AlertRuntime, AlertsService } from "@/services/alerts/AlertsService"
 import { AlertDestinationsService } from "@/services/alerts/AlertDestinationsService"
+import { AlertReadModelsService } from "@/services/alerts/AlertReadModelsService"
 import { HazelOAuthService } from "@/services/auth/HazelOAuthService"
 import { OrgClickHouseSettingsService } from "@/services/org/OrgClickHouseSettingsService"
 import { OrgMembersService } from "@/services/org/OrgMembersService"
@@ -129,6 +130,9 @@ const makeHarness = () => {
 			Layer.mergeAll(envLive, testDb.layer, runtimeLive, hazelOAuthLive, emailLive, orgMembersLive),
 		),
 	)
+	const alertReadModelsLive = AlertReadModelsService.layer.pipe(
+		Layer.provide(Layer.mergeAll(testDb.layer, warehouseLive)),
+	)
 	const alertsLive = AlertsService.layer.pipe(
 		Layer.provide(
 			Layer.mergeAll(
@@ -143,6 +147,7 @@ const makeHarness = () => {
 				orgChSettingsLive,
 				investigationsLive,
 				alertDestinationsLive,
+				alertReadModelsLive,
 			),
 		),
 	)
@@ -151,6 +156,7 @@ const makeHarness = () => {
 		AuthService.layer,
 		DashboardPersistenceService.layer,
 		alertDestinationsLive,
+		alertReadModelsLive,
 		alertsLive,
 	).pipe(Layer.provideMerge(Layer.mergeAll(envLive, testDb.layer)))
