+++ apps/api/src/routes/v2/alert-deliveries.http.ts
@@ -4,7 +4,7 @@ import { CurrentTenant } from "@maple/domain/http"
 import type { V2AlertDelivery } from "@maple/domain/http/v2"
 import { MapleApiV2, paginateOffsetQuery, timestamp, timestampOrNull } from "@maple/domain/http/v2"
 import { Effect } from "effect"
-import { AlertsService } from "@/services/alerts/AlertsService"
+import { AlertReadModelsService } from "@/services/alerts/AlertReadModelsService"
 import { mapAlertError } from "./alerts-error-map"
 
 const toV2Delivery = (doc: AlertDeliveryEventDocument): V2AlertDelivery => ({
@@ -29,12 +29,12 @@ const toV2Delivery = (doc: AlertDeliveryEventDocument): V2AlertDelivery => ({
 
 export const HttpV2AlertDeliveriesLive = HttpApiBuilder.group(MapleApiV2, "alertDeliveries", (handlers) =>
 	Effect.gen(function* () {
-		const alerts = yield* AlertsService
+		const readModels = yield* AlertReadModelsService
 		return handlers.handle("list", ({ query }) =>
 			Effect.gen(function* () {
 				const tenant = yield* CurrentTenant.Context
 				const page = yield* paginateOffsetQuery(query, ({ limit, offset }) =>
-					alerts.listDeliveryEvents(tenant.orgId, { limit, offset }).pipe(
+					readModels.listDeliveryEvents(tenant.orgId, { limit, offset }).pipe(
 						mapAlertError("delivery_list"),
 						Effect.map((response) => response.events.map(toV2Delivery)),
 					),
